Moving PutIntoShadows & ReturnCardToDeck to MoveCardEventGenerator
