+++ server/game/GameActions/ReturnCardToDeck.js
@@ -1,7 +1,6 @@
 import Message from '../Message.js';
 import GameAction from './GameAction.js';
-import LeavePlay from './LeavePlay.js';
-import PlaceCard from './PlaceCard.js';
+import MoveCardEventGenerator from './MoveCardEventGenerator.js';
 
 class ReturnCardToDeck extends GameAction {
     constructor() {
@@ -16,36 +15,16 @@ class ReturnCardToDeck extends GameAction {
     }
 
     canChangeGameState({ card }) {
-        if (card.location === 'play area' && !LeavePlay.allow({ card })) {
-            return false;
-        }
-
         return card.location !== 'draw deck';
     }
 
     createEvent({ card, allowSave = true, bottom = false, orderable }) {
-        let params = {
-            card: card,
-            allowSave: allowSave,
-            bottom: bottom,
-            snapshotName: 'cardStateWhenMoved'
-        };
-        const returnEvent = this.event('onCardReturnedToDeck', params, (event) => {
-            event.thenAttachEvent(
-                PlaceCard.createEvent({
-                    card: event.card,
-                    location: 'draw deck',
-                    bottom,
-                    orderable
-                })
-            );
+        return MoveCardEventGenerator.createReturnCardToDeckEvent({
+            card,
+            allowSave,
+            bottom,
+            orderable
         });
-
-        if (card.location === 'play area') {
-            return this.atomic(returnEvent, LeavePlay.createEvent({ card, allowSave }));
-        }
-
-        return returnEvent;
     }
 }
